Add the base working version
I have worked on a base working version at this point. My next steps will be to improve this further. The current version uses /data_collector/monitor.py as the primary execution. It will handle launching the node.js server on startup, and then fetch the data in real time. /server/index.js is the node.js server and manages the data between the python backend and the node.js front end. index.html is the front in that is served up, and using scripts.js to handle the incoming data and display it to index.html.